DELETEs
Corrected Sql paths and routes, confirmation after clicking in DELETE buton in campaign and user

+++ MacacosMazmorrasMVC/Controllers/CampaignController.cs
@@ -84,11 +84,14 @@ namespace MacacosMazmorrasMVC.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteCampaign(int campaignId)
         {
-            // Delete the campaign from the database based on the ID
-            campaignDAL.DeleteCampaign(campaignId);
+            if (Request.Form["confirmed"] == "true")
+            {
+                campaignDAL.DeleteCampaign(campaignId);
 
-            // Redirect to a success page or any other appropriate action
-            return RedirectToAction("Index", "Campaign");
+                return RedirectToAction("Index", "Campaign");
+            }
+
+            return NoContent();
         }
 
         public IActionResult CamapignWithSesion(int campaignId)
